fixing type annotations
